Added more log statements with more descriptive logging. Added logging to files, minor functionality changes.

+++ src/main/java/com/example/InventorySystem/controllers/InventoryController.java
@@ -18,37 +18,43 @@ import java.util.Optional;
 @RequestMapping("/inventory")
 public class InventoryController {
 
-    private final Logger logger = LoggerFactory.getLogger(InventoryController.class);
+    private final Logger log = LoggerFactory.getLogger(InventoryController.class);
 
     @Autowired
     InventoryServiceImpl inventoryService;
 
     @GetMapping("/")
     public String showAllInventory(Model model) {
+        log.info("Get inventory list");
         try {
             List<Inventory> inventories = inventoryService.getAll();
+            log.info("Inventory list size: {}", inventories.size());
             model.addAttribute("inventories", inventories);
             return "show-all-inventory";
         } catch (Exception e) {
-            logger.error("Error occurred while fetching all inventories", e);
+            log.error("Error occurred while fetching all inventories", e);
             return "error";
         }
     }
 
     @GetMapping("/{inventoryId}")
     public String showOneInventory(@PathVariable("inventoryId") int inventoryId, Model model) {
-        try {
-            Optional<Inventory> inventory = inventoryService.getInventoryById(inventoryId);
+        log.info("Get inventory with id: {}", inventoryId);
+        Optional<Inventory> inventory = inventoryService.getInventoryById(inventoryId);
+
+        if(inventory.isPresent()) {
             model.addAttribute("inventory", inventory);
+            log.info("Received inventory with inventory number: {}", inventory.get().getInventoryNumber());
             return "show-one-inventory";
-        } catch (Exception e) {
-            logger.error("Error occurred while fetching inventory with ID: {}", inventoryId, e);
+        } else {
+            log.info("Inventory with id: {} not found", inventoryId);
             return "error";
         }
     }
 
     @GetMapping("/create")
     public String createInventoryForm(Model model) {
+        log.info("Creating new inventory");
         model.addAttribute("inventory", new Inventory());
         return "create-inventory";
     }
@@ -56,6 +62,7 @@ public class InventoryController {
     @PostMapping("/create")
     public String createInventory(@Valid Inventory inventory, BindingResult result) {
         if (result.hasErrors()) {
+            log.error("Failed creating a new inventory, error: {}", result);
             return "create-inventory";
         }
 
@@ -65,10 +72,12 @@ public class InventoryController {
             temp.setInventoryNumber(inventory.getInventoryNumber());
             temp.setRoom(inventory.getRoom());
             temp.setCabinet(inventory.getCabinet());
+
             inventoryService.createInventory(temp);
+            log.info("Inventory created successfully: {}", temp);
             return "redirect:/inventory/";
         } catch (Exception e) {
-            logger.error("Error occurred while creating inventory", e);
+            log.error("Error occurred while creating inventory", e);
             return "error";
         }
     }
@@ -86,7 +95,7 @@ public class InventoryController {
                 return "redirect:/inventory/";
             }
         } catch (Exception e) {
-            logger.error("Error occurred while preparing update form for inventory with ID: {}", inventoryId, e);
+            log.error("Error occurred while preparing update form for inventory with ID: {}", inventoryId, e);
             return "error";
         }
     }
@@ -94,27 +103,42 @@ public class InventoryController {
     @PostMapping("/update/{inventoryId}")
     public String updateInventoryById(@PathVariable("inventoryId") int inventoryId, @Valid Inventory inventory,
                                       BindingResult result) {
+
+        if (inventoryId <= 0) {
+            log.warn("The id can only be positive, negative value provided: {}", inventoryId);
+            return "inventory-update-page";
+        }
         if (result.hasErrors()) {
+            log.warn("Error updating inventory: {}", result.getAllErrors());
             return "inventory-update-page";
         }
         try {
             inventoryService.updateInventoryById(inventoryId, inventory);
+            log.info("Inventory with id: {} has been updated", inventoryId);
             return "redirect:/inventory/{inventoryId}";
         } catch (Exception e) {
-            logger.error("Error occurred while updating inventory with ID: {}", inventoryId, e);
+            log.error("Error occurred while updating inventory with ID: {}", inventoryId, e);
             return "error";
         }
     }
 
     @PostMapping("/delete/{inventoryId}")
     public String deleteInventoryById(@PathVariable("inventoryId") int inventoryId) {
-        try {
-            inventoryService.deleteInventoryById(inventoryId);
-            return "redirect:/inventory/";
-        } catch (Exception e) {
-            logger.error("Error occurred while deleting inventory with ID: {}", inventoryId, e);
-            return "error";
+
+        if (inventoryId <= 0) {
+            log.warn("The id can only be positive, negative value provided: {}", inventoryId);
+            return "show-one-inventory"; //TODO check if endpoint=correct
         }
+        Optional<Inventory> inventory = inventoryService.getInventoryById(inventoryId);
+        if (!inventory.isPresent()) {
+            log.warn("Inventory with id: {} for delete is not found", inventoryId);
+            return "show-one-inventory"; //TODO same here
+        }
+
+        inventoryService.deleteInventoryById(inventoryId);
+        log.info("Inventory with id: {} is deleted", inventoryId);
+        return "redirect:/inventory/";
     }
 }
+
